feat: Add `mean` function & feat(compiler): Added compiler tokenizing and rewrote the compiler
feat: Add `mean` function and multi-param feature of functions
feat(compiler): Added compiler tokenizing and rewrote the compiler

In order to help compiler better understanding multi-param, the tokenizing feature is needed.
Tokenizing makes it more convenient to calculate the formula,
because the structure of a tokenized formula is more clearer than a non-tokenized one,
and in this way, we can process the formula much easilier.